Fix some plugin issues
